Word Frequency in Classic Novels
In this Project, we scrape a novel from the website Project Gutenberg (which contains a large corpus of books) using the Python package requests. Then we extract words from this web data using BeautifulSoup. Finally, we'll dive into analyzing the distribution of words using the Natural Language ToolKit (nltk).

The natural language processing tools used here apply to much of the data that data scientists encounter as a vast proportion of the world's data is unstructured data and includes a great deal of text. To complete this Project, we need to know how to import web data into Python and how to work with natural language text.